Add Statistic Information.

+++ ImageTool/FormImgTool.cs
@@ -58,11 +58,11 @@ namespace ImageTool
                     file.Write(Environment.NewLine);
                 }
 
-                double radiusStdEv = Utils.Math.StdEv(f.Rounds.Select(x => x.MaxLenLine.Length).ToList());
+                double radiusStdEv = Utils.Math.StdEv(f.Rounds.Select(x => (double)x.MaxLenLine.Length).ToList());
                 file.Write(string.Format("StdEv of Radius: {0}", radiusStdEv));
                 file.Write(Environment.NewLine);
 
-                double weightStdEv = Utils.Math.StdEv(f.Rounds.Select(x => x.Weight).ToList());
+                double weightStdEv = Utils.Math.StdEv(f.Rounds.Select(x => (double)x.Weight).ToList());
                 file.Write(string.Format("StdEv of Weight: {0}", weightStdEv));
                 file.Write(Environment.NewLine);
             }
